test(cli): address smoke-test review feedback
- hub-crud: pass parsed objects to canonicalEqual so the pulled/stored
  comparison is key-order independent instead of degrading to byte equality
- global-setup: guard compose down in the failure path so the original
  'CalmHub not ready' error is not masked by a teardown failure
- hub-ready: bound each readiness poll with AbortSignal.timeout so a stalled
  connection cannot hang past the deadline
- cli-runner: clean up the temp dir (and any stray tarball) if pack/install
  fails before a CliInstall is returned to the caller
- extract writeJson/readJson/patchJson into src/test_helpers/json-file.ts,
  shared by cli.e2e.spec.ts and the smoke fixtures harness
